# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Palace Jewelry API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName: string;
  lastName: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  createdAt: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  originalPrice?: string | null;
  imageUrl: string;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  isNewArrival: boolean;
  isTrending: boolean;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  weight?: string | null;
  createdAt: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CartItem {
  id: number;
  userId: number;
  productId: number;
  quantity: number;
  product: Product;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface NewsletterSubscribeBody {
  email: string;
}

export type ListProductsParams = {
  categoryId?: number;
  isNewArrival?: boolean;
  isTrending?: boolean;
};