# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Palace Jewelry API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  firstName: zod.string(),
  lastName: zod.string(),
});

/**
 * @summary Login with email and password
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  firstName: zod.string(),
  lastName: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user session
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  firstName: zod.string(),
  lastName: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all products
 */
export const ListProductsQueryParams = zod.object({
  categoryId: zod.coerce.number().optional(),
  isNewArrival: zod.coerce.boolean().optional(),
  isTrending: zod.coerce.boolean().optional(),
});

export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  price: zod.string(),
  originalPrice: zod.string().nullish(),
  imageUrl: zod.string(),
  categoryId: zod.number().nullish(),
  categoryName: zod.string().nullish(),
  isNewArrival: zod.boolean(),
  isTrending: zod.boolean(),
  badge: zod.string().nullish(),
  weight: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Get a product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  price: zod.string(),
  originalPrice: zod.string().nullish(),
  imageUrl: zod.string(),
  categoryId: zod.number().nullish(),
  categoryName: zod.string().nullish(),
  isNewArrival: zod.boolean(),
  isTrending: zod.boolean(),
  badge: zod.string().nullish(),
  weight: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  imageUrl: zod.string().nullish(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get current user's cart
 */
export const GetCartResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  productId: zod.number(),
  quantity: zod.number(),
  product: zod.object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string().nullish(),
    price: zod.string(),
    originalPrice: zod.string().nullish(),
    imageUrl: zod.string(),
    categoryId: zod.number().nullish(),
    categoryName: zod.string().nullish(),
    isNewArrival: zod.boolean(),
    isTrending: zod.boolean(),
    badge: zod.string().nullish(),
    weight: zod.string().nullish(),
    createdAt: zod.coerce.date(),
  }),
});
export const GetCartResponse = zod.array(GetCartResponseItem);

/**
 * @summary Add item to cart
 */
export const AddToCartBody = zod.object({
  productId: zod.number(),
  quantity: zod.number(),
});

/**
 * @summary Update cart item quantity
 */
export const UpdateCartItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCartItemBody = zod.object({
  quantity: zod.number(),
});

export const UpdateCartItemResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  productId: zod.number(),
  quantity: zod.number(),
  product: zod.object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string().nullish(),
    price: zod.string(),
    originalPrice: zod.string().nullish(),
    imageUrl: zod.string(),
    categoryId: zod.number().nullish(),
    categoryName: zod.string().nullish(),
    isNewArrival: zod.boolean(),
    isTrending: zod.boolean(),
    badge: zod.string().nullish(),
    weight: zod.string().nullish(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Remove item from cart
 */
export const RemoveCartItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Subscribe to newsletter
 */
export const SubscribeNewsletterBody = zod.object({
  email: zod.string().email(),
});