# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Palace Jewelry API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  originalPrice?: string | null;
  imageUrl: string;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  isNewArrival: boolean;
  isTrending: boolean;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  weight?: string | null;
  createdAt: Date;
}